requirements.txt
click==6.7
Flask==1.0.2
flask-table==0.5.0
Flask-Babel==0.12.2
itsdangerous==0.24
Jinja2>=2.10.1
MarkupSafe==1.0
Werkzeug==0.14.1
pandas==0.23.4
azure-storage==0.36.0